+++ web/src/modules/travel-authorizations/pages/my-travel-requests/MyTravelRequestsRequestEditPage.vue
@@ -1,11 +1,11 @@
 <template>
   <v-container>
-    <TravelDeskTravelRequestCreateCard :travel-authorization-id="travelAuthorizationId" />
+    <TravelDeskTravelRequestEditCard :travel-authorization-id="travelAuthorizationId" />
   </v-container>
 </template>
 
 <script setup>
-import TravelDeskTravelRequestCreateCard from "@/components/travel-desk-travel-requests/TravelDeskTravelRequestCreateCard.vue"
+import TravelDeskTravelRequestEditCard from "@/components/travel-desk-travel-requests/TravelDeskTravelRequestEditCard.vue"
 
 defineProps({
   travelAuthorizationId: {
